+++ Plugin.cs
@@ -1,9 +1,5 @@
 using BepInEx;
 using BepInEx.Logging;
-using SigilADay_julianperge.lib;
-using UnityEngine;
-
-// using UnityEngine;
 
 namespace SigilADay_julianperge
 {
@@ -13,19 +9,19 @@ namespace SigilADay_julianperge
 	{
 		public const string PluginGuid = "julianperge.inscryption.sigiladay";
 		private const string PluginName = "SigilADay_julianperge";
-		private const string PluginVersion = "1.2.0";
+		private const string PluginVersion = "1.3.0";
 
 		internal static ManualLogSource Log;
 
 		private void Awake()
 		{
 			Log = base.Logger;
-			
-			
+
+			AddCannibal();
 			AddExcavator();
 			AddFlightOfTheValkyrie();
 			AddMiasma();
-			AddNest();;
+			AddNest();
 			AddSplit();
 		}
 	}