+++ sigils/Valkyrie.cs
@@ -13,8 +13,8 @@ namespace SigilADay_julianperge
 	{
 		private void AddNorseWarrior()
 		{
-			Texture2D defaultTexture = new Texture2D(2, 2);
-			defaultTexture.LoadImage(SigilADay_julianperge.Properties.Resources.card_norse_warrior);
+			Texture2D defaultTexture =
+				SigilUtils.LoadTextureFromResource(SigilADay_julianperge.Properties.Resources.card_norse_warrior);
 
 			const string name = "NorseWarrior";
 			const string displayName = "Norse Warrior";
@@ -31,27 +31,19 @@ namespace SigilADay_julianperge
 			AddNorseWarrior();
 
 			const string rulebookName = "Flight of the Valkyrie";
-			const string description =
+			const string rulebookDescription =
 				"When [creature] is played, fill all open slots on your side of the field with 1/1 Norse Warriors.";
 
 			// setup ability
-			AbilityInfo info = SigilUtils.CreateInfoWithDefaultSettings(rulebookName, description, 3);
-
-			List<DialogueEvent.Line> lines = new List<DialogueEvent.Line>();
-			DialogueEvent.Line line = new DialogueEvent.Line
-			{
-				text = "When this creature is played, fill all open slots on your side of the field with 1/1 Norse Warriors."
-			};
-			lines.Add(line);
-			info.abilityLearnedDialogue = new DialogueEvent.LineSet(lines);
+			AbilityInfo info = SigilUtils.CreateInfoWithDefaultSettings(rulebookName, rulebookDescription, true, 3);
 
 			// get and load artwork
-			Texture2D sigilTex = new Texture2D(2, 2);
-			sigilTex.LoadImage(SigilADay_julianperge.Properties.Resources.ability_fotv);
+			Texture2D sigilTex = SigilUtils.LoadTextureFromResource(SigilADay_julianperge.Properties.Resources.ability_fotv);
 
-			// set ability to behavior class
-			var abIds = SigilUtils.GetAbilityId(info.rulebookName);
-			NewAbility newAbility = new NewAbility(info, typeof(FlightOfTheValkyrie), sigilTex, abIds);
+			var abId = SigilUtils.GetAbilityId(info.rulebookName);
+			
+			NewAbility newAbility = new NewAbility(info, typeof(FlightOfTheValkyrie), sigilTex, abId);
+			
 			FlightOfTheValkyrie.ability = newAbility.ability;
 
 			return newAbility;